fix: add same check used for creating BACKUP_AGENT_CREDENTIALS for its deletion. deleting an existing service crashed when the credentials were not created, because the deletion of the credentials failed.

+++ osb-service/src/main/java/de/evoila/cf/cpi/existing/MongoDbExistingServiceFactory.java
@@ -69,8 +69,11 @@ public class MongoDbExistingServiceFactory extends ExistingServiceFactory {
 
         mongoDBCustomImplementation.deleteDatabase(mongoDbService, MongoDBUtils.dbName(serviceInstance.getId()));
 
+        if(existingEndpointBean.getBackupCredentials() != null) {
+            credentialStore.deleteCredentials(serviceInstance, DefaultCredentialConstants.BACKUP_AGENT_CREDENTIALS);
+        }
+
         credentialStore.deleteCredentials(serviceInstance, CredentialConstants.ROOT_CREDENTIALS);
-        credentialStore.deleteCredentials(serviceInstance, DefaultCredentialConstants.BACKUP_AGENT_CREDENTIALS);
         credentialStore.deleteCredentials(serviceInstance, DefaultCredentialConstants.BACKUP_CREDENTIALS);
     }
